feat: improve authentication handling and fix hydration issues
- Add NEXT_PUBLIC_FRONTEND_BASE_URL to environment configs
- Refactor app/page.tsx to use AuthGuard component for cleaner auth logic
- Add immediate client-side hydration to useAuthStore
- Create AuthGuard component with dynamic import to avoid SSR issues
- Remove complex authentication logic from main page component
