+++ .env.dev
@@ -2,6 +2,7 @@
 # This file is used when running "npm run dev"
 SWAGGER_URL=http://158.179.175.134:8080/v3/api-docs/Motimo%20API%20List
 NEXT_PUBLIC_API_URL=""
+NEXT_PUBLIC_FRONTEND_BASE_URL=http://localhost:3000
 
 # Default to remote API server
 USE_LOCAL_API=false
