+++ .env.local
@@ -2,6 +2,7 @@
 # This file is used when running "npm run local"
 SWAGGER_URL=http://localhost:8080/v3/api-docs/Motimo%20API%20List
 NEXT_PUBLIC_API_URL=""
+NEXT_PUBLIC_FRONTEND_BASE_URL=http://localhost:3000
 
 # Use local API server when running "npm run local"
 USE_LOCAL_API=true
